Add debugging and profiling tools for C binaries.
Adds support for spawning a gdbgui debugger as well as the ability
to run profiling on solver binaries with gprof.

To enable debugging, run solver.compile with `debug=True`. This
adds the `-g` flag into the `gcc` commands that do the compilation,
enabling debugging information to be injected into the resulting
binaries.

To start the debugger, run `solver.run_debugger()`. This will open
a new gdbgui session at `http://127.0.0.1:5000`. Use the result of
the compile() statement to get the build directory of the compiled
binary and place it in the path in gdbgui, followed by the name
of the executable, currently `ssa_sdpd`. For example, if your build
directory is `/tmp/spatialpy_build_0q9u0_lg` then you would put the
path `/tmp/spatialpy_build_0q9u0_lg/ssa_sdpd` into the path field
in gdbgui and then click "Load Binary."

If your solver has not been compiled and you rely on run() to
compile the binary implicitly, passing `debug=True` to run() will
run `compile()` with `debug=True`.

To enable profiling, you must invoke compile() and/or run() with
`profile=True`. The binary must have been compiled with `profile=True`.
If you invoke `run()` with `profile=True` on a binary that was
compiled with `profile=False` it will not work! Anyway, once the run
completes the gprof report for each trajectory will be printed.